feat: Extract templates from components
Clean up imports and export EmailVerification, ResetPassword, and ResendEmailVerification components
