+++ src/authport/authport.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit, TemplateRef, Input} from '@angular/core';
+import { Component, OnInit, TemplateRef, Input } from '@angular/core';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { Stormpath, LoginService } from '../stormpath/stormpath.service';
