+++ src/register/register.component.html
@@ -1,0 +1,23 @@
+<template #defaultTemplate>
+  <form *ngIf="!registered" (ngSubmit)="onSubmit()" class="form-horizontal">
+    <div class="form-group" *ngFor="let field of model?.form?.fields">
+      <label [attr.for]="field.name" class="col-sm-4 control-label">{{field.label}}</label>
+      <div class="col-sm-8">
+        <input class="form-control" [name]="field.name" [id]="field.name" [type]="field.type"
+               [(ngModel)]="formModel[field.name]" [placeholder]="field.placeholder" [disabled]="creating" [required]="field.required">
+      </div>
+    </div>
+    <div *ngIf="error" class="alert alert-danger">{{error}}</div>
+    <button type="submit" class="btn btn-primary">Register</button>
+  </form>
+  <p *ngIf="unverified" class="alert alert-success">
+    Your account has been created and requires verification.
+    Please check your email for a verification link.
+  </p>
+  <p class="alert alert-success" *ngIf="canLogin">
+    Your account has been created, you may now log in.
+  </p>
+</template>
+<template
+  [ngTemplateOutlet]="customTemplate || defaultTemplate">
+</template>
