+++ src/reset-password/reset-password.component.html
@@ -1,0 +1,40 @@
+<template #defaultTemplate>
+<div class="row">
+  <div class="col-sm-offset-4 col-xs-12 col-sm-4">
+    <p *ngIf="verifying" class="alert alert-warning text-center">We are verifying your password reset link</p>
+    <p class="alert alert-success" *ngIf="reset">Your new password has been set, you may now login.</p>
+    <div *ngIf="verificationFailed" class="alert alert-danger">
+      This password reset link is not valid, please request a new reset link.
+    </div>
+  </div>
+</div>
+<div class="row">
+  <div class="col-xs-12">
+    <form class="form-horizontal" *ngIf="verified && !reset" (ngSubmit)="onSubmit()">
+      <div class="form-group">
+        <label for="spUsername" class="col-xs-12 col-sm-4 control-label">New Password</label>
+        <div class="col-xs-12 col-sm-4">
+          <input class="form-control" id="spUsername" [(ngModel)]="formModel.password"
+                 placeholder="New Password" type="password" [disabled]="posting">
+        </div>
+      </div>
+      <div class="form-group">
+        <label for="spPassword" class="col-xs-12 col-sm-4 control-label">Confirm New Password</label>
+        <div class="col-xs-12 col-sm-4">
+          <input class="form-control" id="spPassword" [(ngModel)]="formModel.confirmPassword"
+                 placeholder="Confirm New Password" type="password" [disabled]="posting">
+        </div>
+      </div>
+      <div class="form-group">
+        <div class="col-sm-offset-4 col-sm-4">
+          <p class="alert alert-danger" *ngIf="error">{{error}}</p>
+          <button type="submit" class="btn btn-primary" [disabled]="posting">Set New Password</button>
+        </div>
+      </div>
+    </form>
+  </div>
+</div>
+</template>
+<template
+  [ngTemplateOutlet]="customTemplate || defaultTemplate">
+</template>
