+++ test/stormpath/stormpath.http.spec.ts
@@ -1,11 +1,7 @@
 import { TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
-import { StormpathModule } from '../../src/stormpath.module';
-import { Stormpath } from '../../src/stormpath/stormpath.service';
+import { Account, StormpathModule, Stormpath, StormpathConfiguration, StormpathHttp } from '../../src';
 import { MockBackend } from '@angular/http/testing';
 import { Response, ResponseOptions, BaseRequestOptions, Http, ConnectionBackend } from '@angular/http';
-import { Account } from '../../src/shared/account';
-import { StormpathConfiguration } from '../../src/stormpath/stormpath.config';
-import { StormpathHttp, CurrentDomain } from '../../src/stormpath/stormpath.http';
 
 describe('StormpathHttp', () => {
 
@@ -142,7 +138,12 @@ describe('StormpathHttp', () => {
           });
           c.mockRespond(new Response(response));
         });
-        stormpath.register({email: 'bar@testmail.stormpath.com', password: 'foo', surname: 'Bar', givenName: 'Foo'}).subscribe((response) => {
+        stormpath.register({
+          email: 'bar@testmail.stormpath.com',
+          password: 'foo',
+          surname: 'Bar',
+          givenName: 'Foo'
+        }).subscribe((response) => {
           account = response;
         });
         tick();
