+++ test/stormpath/stormpath.service.spec.ts
@@ -1,11 +1,7 @@
 import { TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
-import { StormpathModule } from '../../src/stormpath.module';
-import { Stormpath } from '../../src/stormpath/stormpath.service';
+import { Account, Stormpath, StormpathModule } from '../../src';
 import { MockBackend } from '@angular/http/testing';
 import { Response, ResponseOptions, BaseRequestOptions, Http, ConnectionBackend } from '@angular/http';
-import { Account } from '../../src/shared/account';
-import { Observable } from 'rxjs';
-import { BaseStormpathAccount } from '../../src/shared/account';
 
 describe('StormpathService', () => {
